+++ src/actions/settings/loadUserSettingsFromStorage.js
@@ -1,0 +1,14 @@
+import storageStub from '../../storage/storageStub';
+
+const storage = localStorage || storageStub;
+
+export default () => (dispatch) => new Promise((resolve) => {
+    dispatch({
+        type: 'SAVE_USER_SETTINGS',
+        settings: {
+            displayLanguage: storage.getItem('displayLanguage') || 'de-DE'
+        }
+    });
+
+    resolve();
+});