+++ src/containers/stages/header/SearchContainer.js
@@ -2,14 +2,17 @@ import { connect } from 'react-redux';
 import Items from '../../../components/stages/header/Search';
 import searchInformationItems from '../../../actions/items/searchInformationItems';
 import { bindActionCreators } from 'redux';
+import getTranslation from '../../../helpers/getTranslation';
 
 const mapStateToProps = (state) => {
     const { activeNavigation } = state.navigation;
 
     return {
         activeNavigation,
-        searchButtonLabel: 'suchen',
-        searchTextPlaceholder: 'Suchbegriff'
+        translations: {
+            searchButtonLabel: getTranslation(state, 'SEARCH_SEARCH_BUTTON_LABEL'),
+            searchTextPlaceholder: getTranslation(state, 'SEARCH_SEARCH_TEXT_PLACEHOLDER')
+        }
     };
 };