+++ src/reducers/user.js
@@ -1,13 +1,21 @@
 const defaultState = {
     currentUserId: '',
-    guestUserId: ''
+    guestUserId: '',
+    settings: {
+        displayLanguage: 'de-DE'
+    }
 };
 
 export default (state = defaultState, action) => {
     const getNewState = (newState) => Object.assign({}, state, newState);
 
-    const { userId } = action;
+    const { settings, userId } = action;
     switch (action.type) {
+        case 'SAVE_USER_SETTINGS': {
+            return getNewState({
+                settings
+            });
+        }
         case 'SET_CURRENT_USER_ID':
             return getNewState({
                 currentUserId: userId