+++ src/repositories/settings/getUserSettingsMockRepository.js
@@ -1,0 +1,5 @@
+import userSettingsMock from '../../__data__/user-settings-mock';
+
+export default (currentUserId) => new Promise((resolve) => {
+    resolve(userSettingsMock[currentUserId] || {});
+});